+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Image Uploader now has translated messages.
+
 ## [3.11.0] - 2019-05-28
 
 ## [3.10.5] - 2019-05-28